Add gyp to python packages.
This is normally needed for for building packages related to Chromium, but as
Chromium itself is using a patched version of gyp, it's not used there. However
for testing purposes and for packaging chromium and related stuff it's very
useful to have the upstream version available.

In addition, it might be useful for someone out there, who knows?