libdrm: Update package to version 2.4.39.
This is needed in order to properly build the new Mesa version.
We now enable nouveau/radeon/intel explicitly to get a build error in case we're
missing a dependency.

In addition, you may have noticed, that we no longer need the -experimental-api,
because these drivers are out of staging.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>