steam: Base package on Steam internal packages.
So far we used the steam binary from the Debian package. This isn't really
needed as the binary as well as all those scripts are mostly used for
bootstrapping only.

The new approach is fetching _all_ required files directly from
http://media.steampowered.com/ and create a fake package and .installed file.

I'm not yet 100% clear what the .installed file _should_ contain, but after some
debugging I found out that it mostly contains paths and the corresponding file
sizes. The format in detail is like this:

<path>,<file_size>\n

If the path is a directory, file size is -1.

But it doesn't seem to be the whole truth as there are library calls to strchr()
even after the file size. In addition there are even strchr() calls operating on
':' so it might be useful if we would have an actual example file.

In addition, the new appreach now fetches all resource files needed for Steam,
so we could possibly avoid updating in the first place and already have a
readily bootstrapped Steam in the derivation output.

So, while working on packaging Steam it came to my attention (a bit too late if
you ask me), that @cpages is also working on packaging Steam. He began way
earlier (11 days ago, so about week before I started off) and might have better
chances of success because he actually has closed beta access.

Well, but the more experts we have on a particular package the better. It's just
a shame of the duplicate work, so sorry @cpages for not noticing your branch.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>