steam: Add ld preloader for access control.
My nix store is quite large and it takes aaageeees to read every damn item
within /nix/store. And no, this is not a "hidden spyware functionality" of Steam
but comes from the win32/posix compatibility library they're using, which is
looking up file names case-insensitive.

Well, the wrapper currently only allows access to the output path of the
derivation and $HOME/Steam, but I now decided to go nuclear and my next step
will be to implement seccomp BPF on Steam.

Of course we could reuse the implementation or at least the parts/experience
from it in nix-daemon as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>