steam: Add basic seccomp BPF sandbox.
This really is quite basic and just gets rid of breakpad and a few more
updater-related functionalitios. Hope I didn't get banned for sending in so many
crash reports during the course of debugging.

The BPF program currently _only_ does filtering and doesn't implement traps so
we can get rid of the LD_PRELOAD stuff as well.

But as promised, we went nuclear, so Steam is not even able to segfault anymore
because it doesn't have permissions to do anything signal-related.

Now, wouldn't it be nice if we can write seccomp filters using Nix expressions?

Signed-off-by: aszlig <aszlig@redmoonstudios.org>