chromium: Update all channels to latest versions.
This introduces version 31.0.1650.57 as the new version for the stable
channel.

Overview of the updated channels:

stable: 30.0.1599.114 -> 31.0.1650.57
beta:   31.0.1650.34  -> 32.0.1700.19
dev:    32.0.1671.3   -> 33.0.1712.4

This drops the sandbox_userns_30.patch as version 30 is no longer
stable. In addition, we had to patch out some references to /usr/bin/gcc
in the bundled WebKit sources.

Builds are passing and running fine.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>