generic/setup.sh: Make env hooks POSIX-compliant.
Unfortunately, when calling the list (or better: the space-separated
string of strings) of env hooks we'll separate by IFS as well, so env
hooks containing a space in the name won't work properly.

As we do not have any env hooks that have a space in their name, I think
this is not a problem.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>