qemu: Update to new upstream version 2.2.0.
The patch for CVE-2014-0150 has never _really_ applied because it was in
invalid patch format (look at the <at>'s instead of "@"). But because it
was applied with patch instead of git apply, it never bailed out with an
error, because patch treated everything after the header as comments and
thus entirely skipped the hunk.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>