nixos/tests/keymap: Make screenshots on timeouts
This is a temporary workaround, so we see what's going on whenever the
Hydra build fails. It touches the file $out/nix-support/failed, so that
the Hydra job is marked as failed but the build still succeeds so we get
an actual output including a screen shot of the failure.

Unfortunately this has the drawback, that we can't easily restart such a
failed build unless we remove the store path.

Signed-off-by: aszlig <aszlig@nix.build>