qemu_test: Add video encoder
This tool converts the intermediate format that we now get from QEMU
when running with --nixos-test outfile into any video format that can be
played by most video players, particularly web browsers when watching
the videos on Hydra.

The video format to be used is determined by the output file name, so
it's easy to change it to a different format rather than WebM (which is
the format this was designed for) by simply changing the file name.

I also have to give credit to "kepstin" from the #ffmpeg channel on
freenode for helping with the rescaling of the presentation timestamp.

Signed-off-by: aszlig <aszlig@nix.build>