nixos/testing: Encode videos by default
This now encodes the raw video frames for every single VM test in our
internal format. Due to compression the average video file size is about
~3M so I think it's a good idea to enable it by default so we can see
post-mortem what went wrong.

Instead of killing the machines with the SIGKILL signal during cleanup,
we now kill it with SIGTERM because otherwise the atexit() handler
doesn't run and we get a video file without the gzip buffer flushed.

I'm also adding a helper attribute called 'videos' to every test, so
that all these videos can be encoded on Hydra. The derivation propagates
the build products from the actual test runner and also adds videos to
the Hydra build products.

Signed-off-by: aszlig <aszlig@nix.build>