qemu_test: Move video encoder into own derivation
If we only want to run tests without actually encoding the videos, it
doesn't really make sense to pull in the FFmpeg dependency.

When debugging tests we can still either use the encoder manually or
simply append the .videos attribute in order to get the encoded videos
(which then of course will pull in the FFmpeg dependency).

Signed-off-by: aszlig <aszlig@nix.build>